Fix a bunch of compiler warnings that show up in windows build

+++ src/builtin_functions/cos.cpp
@@ -9,7 +9,7 @@ namespace cos_function {
     {
         float input = to_float(caller->input(0));
         // Convert input from 0..1 to 0..2pi
-        as_float(caller) = cos(input * 2 * M_PI);
+        as_float(caller) = cos(float(input * 2 * M_PI));
     }
 
     void feedback_evaluate(Term* caller)
@@ -26,7 +26,7 @@ namespace cos_function {
         float result = std::acos(desired);
 
         // Map result from radians into range of 0..1
-        as_float(caller) = result / (2 * M_PI);
+        as_float(caller) = result / float(2 * M_PI);
     }
 
     void setup(Branch& kernel)
