+++ src/builtin_functions/sin.cpp
@@ -10,7 +10,7 @@ namespace sin_function {
         float input = to_float(caller->input(0));
 
         // Convert input from 0..1 to 0..2pi
-        as_float(caller) = sin(input * 2 * M_PI);
+        as_float(caller) = sin(float(input * 2 * M_PI));
     }
 
     void feedback_evaluate(Term* caller)
@@ -27,7 +27,7 @@ namespace sin_function {
         float result = std::asin(desired);
 
         // Map result from radians into range of 0..1
-        as_float(caller) = result / (2 * M_PI);
+        as_float(caller) = result / float(2 * M_PI);
     }
 
     void setup(Branch& kernel)
